Move field and methods to base class from TestMySqlConnectorTest
